requirements.txt
about-time==4.2.1
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alive-progress==3.2.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
autograd==1.7.0
bitsandbytes==0.45.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cma==3.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.4.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jstyleson==0.0.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.31.0
natsort==8.4.0
networkx==3.3
ninja==1.11.1.3
nncf==2.15.0
numpy==2.1.3
openai==1.66.3
packaging==24.2
pandas==2.2.3
peft==0.15.0
pillow==11.1.0
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==2.0.0
Pygments==2.19.1
pymoo==0.6.1.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
setuptools==77.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.43.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
trl==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
openai-whisper==20240930
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
youtube-search==2.1.2